pyproject.toml
[project]
name = "finance-mcp"
version = "0.1.0"
description = "Personal Finance MCP Server - Production-ready financial management via MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary",
    "alembic",
    "pydantic>=2",
    "pydantic-settings>=2",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "httpx>=0.27",
]

[project.scripts]
finance-mcp = "app.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
